refactor: 把gorm换成了xorm
1. 重构项目，将gorm换成了xorm
2. 取消200返回时的模板，直接返回内容
3. 添加错误时的错误模板，对错误大致分类编码

+++ model/reqo/message.go
@@ -2,13 +2,13 @@ package reqo
 
 // GetMessage GET "/message" request object
 type GetMessage struct {
-	BeginID   uint  `json:"beginId"`
+	BeginID   int64 `json:"beginId"`
 	BeginTime int64 `json:"beginTime"`
 	EndTime   int64 `json:"endTime"`
 }
 
 // PostMessage POST "/message" request object
 type PostMessage struct {
-	ReceiverID uint   `json:"receiverId"`
+	ReceiverID int64  `json:"receiverId"`
 	Content    string `json:"content"`
 }
