+++ model/reso/user.go
@@ -2,7 +2,7 @@ package reso
 
 // GetUser GET "/user" response object
 type GetUser struct {
-	ID       uint   `json:"id"`
+	ID       int64  `json:"id"`
 	Username string `json:"username"`
 	Gender   int64  `json:"gender"`
 	Age      int64  `json:"age"`
@@ -11,7 +11,7 @@ type GetUser struct {
 
 // PutUser PUT "/user" response object
 type PutUser struct {
-	ID       uint   `json:"id"`
+	ID       int64  `json:"id"`
 	Username string `json:"username"`
 	Gender   int64  `json:"gender"`
 	Age      int64  `json:"age"`
@@ -21,12 +21,12 @@ type PutUser struct {
 // PostUser POST "/user" response object
 type PostUser struct {
 	Username string `json:"username"`
-	ID       uint   `json:"id"`
+	ID       int64  `json:"id"`
 }
 
 // PostLogin POST "/login" response object
 type PostLogin struct {
 	Username string `json:"username"`
-	ID       uint   `json:"id"`
+	ID       int64  `json:"id"`
 	Token    string `json:"token"`
 }
